Adding a common client module and class.
Refactor the gem to have common functinality extracted.
