+++ lib/kubeclient/watch_notice.rb
@@ -1,4 +1,7 @@
 require 'recursive_open_struct'
-
-class WatchNotice <  RecursiveOpenStruct
+module Kubeclient
+  module Common
+    class WatchNotice < RecursiveOpenStruct
+    end
+  end
 end
