+++ test/test_kubeclient.rb
@@ -66,7 +66,7 @@ class KubeClientTest < MiniTest::Test
   end
 
   def test_api
-    stub_request(:get, 'http://localhost:8080/api/')
+    stub_request(:get, 'http://localhost:8080/api')
       .to_return(status: 200, body: open_test_json_file('versions_list.json'))
 
     client = Kubeclient::Client.new 'http://localhost:8080/api/', 'v1beta3'
@@ -75,34 +75,34 @@ class KubeClientTest < MiniTest::Test
   end
 
   def test_api_valid
-    stub_request(:get, 'http://localhost:8080/api/')
+    stub_request(:get, 'http://localhost:8080/api')
       .to_return(status: 200, body: open_test_json_file('versions_list.json'))
 
-    client = Kubeclient::Client.new 'http://localhost:8080/api/', 'v1beta3'
+    client = Kubeclient::Client.new 'http://localhost:8080/api/'
     assert client.api_valid?
   end
 
   def test_api_valid_with_invalid_json
-    stub_request(:get, 'http://localhost:8080/api/')
+    stub_request(:get, 'http://localhost:8080/api')
       .to_return(status: 200, body: '{}')
 
-    client = Kubeclient::Client.new 'http://localhost:8080/api/', 'v1beta3'
+    client = Kubeclient::Client.new 'http://localhost:8080/api/'
     refute client.api_valid?
   end
 
   def test_api_valid_with_bad_endpoint
-    stub_request(:get, 'http://localhost:8080/api/')
+    stub_request(:get, 'http://localhost:8080/api')
       .to_return(status: [404, 'Resource Not Found'])
 
-    client = Kubeclient::Client.new 'http://localhost:8080/api/', 'v1beta3'
+    client = Kubeclient::Client.new 'http://localhost:8080/api/'
     assert_raises(KubeException) { client.api_valid? }
   end
 
   def test_api_valid_with_non_json
-    stub_request(:get, 'http://localhost:8080/api/')
+    stub_request(:get, 'http://localhost:8080/api')
       .to_return(status: 200, body: '<html></html>')
 
-    client = Kubeclient::Client.new 'http://localhost:8080/api/', 'v1beta3'
+    client = Kubeclient::Client.new 'http://localhost:8080/api/'
     assert_raises(JSON::ParserError) { client.api_valid? }
   end
 
@@ -131,7 +131,7 @@ class KubeClientTest < MiniTest::Test
     services = client.get_services
 
     refute_empty(services)
-    assert_instance_of(Kubeclient::EntityList, services)
+    assert_instance_of(Kubeclient::Common::EntityList, services)
     assert_equal('Service', services.kind)
     assert_equal(2, services.size)
     assert_instance_of(Kubeclient::Service, services[0])
@@ -145,7 +145,7 @@ class KubeClientTest < MiniTest::Test
 
     client = Kubeclient::Client.new 'http://localhost:8080/api/', 'v1beta3'
     pods = client.get_pods
-    assert_instance_of(Kubeclient::EntityList, pods)
+    assert_instance_of(Kubeclient::Common::EntityList, pods)
     assert_equal(0, pods.size)
   end
 
@@ -180,12 +180,13 @@ class KubeClientTest < MiniTest::Test
     client = Kubeclient::Client.new 'http://localhost:8080/api/', 'v1beta3'
     result = client.all_entities
     assert_equal(7, result.keys.size)
-    assert_instance_of(Kubeclient::EntityList, result['node'])
-    assert_instance_of(Kubeclient::EntityList, result['service'])
-    assert_instance_of(Kubeclient::EntityList, result['replication_controller'])
-    assert_instance_of(Kubeclient::EntityList, result['pod'])
-    assert_instance_of(Kubeclient::EntityList, result['event'])
-    assert_instance_of(Kubeclient::EntityList, result['namespace'])
+    assert_instance_of(Kubeclient::Common::EntityList, result['node'])
+    assert_instance_of(Kubeclient::Common::EntityList, result['service'])
+    assert_instance_of(Kubeclient::Common::EntityList,
+                       result['replication_controller'])
+    assert_instance_of(Kubeclient::Common::EntityList, result['pod'])
+    assert_instance_of(Kubeclient::Common::EntityList, result['event'])
+    assert_instance_of(Kubeclient::Common::EntityList, result['namespace'])
     assert_instance_of(Kubeclient::Service, result['service'][0])
     assert_instance_of(Kubeclient::Node, result['node'][0])
     assert_instance_of(Kubeclient::Event, result['event'][0])
